requirements.txt
aiohttp
hikari[speedups]
hikari-lightbulb
python-dotenv
uvloop; sys_platform != 'nt'